Move Diagnostic subclasses to Diagnostic object
