+++ compiler/src/dotty/tools/dotc/reporting/diagnostic/Diagnostic.scala
@@ -5,10 +5,12 @@ package diagnostic
 
 import util.SourcePosition
 import core.Contexts.Context
+import config.Settings.Setting
+import interfaces.Diagnostic.{ERROR, INFO, WARNING}
 
 import java.util.Optional
 
-object Diagnostic {
+object Diagnostic:
   val nonSensicalStartTag: String = "<nonsensical>"
   val nonSensicalEndTag: String = "</nonsensical>"
 
@@ -21,7 +23,68 @@ object Diagnostic {
       }
     }
   }
-}
+
+  // `Diagnostics to be consumed by `Reporter` ---------------------- //
+  class Error(
+    msg: Message,
+    pos: SourcePosition
+  ) extends Diagnostic(msg, pos, ERROR)
+
+  /** A sticky error is an error that should not be hidden by backtracking and
+   *  trying some alternative path. Typically, errors issued after catching
+   *  a TypeError exception are sticky.
+   */
+  class StickyError(
+    msg: Message,
+    pos: SourcePosition
+  ) extends Error(msg, pos)
+
+  class Warning(
+    msg: Message,
+    pos: SourcePosition
+  ) extends Diagnostic(msg, pos, WARNING) {
+    def toError: Error = new Error(msg, pos)
+  }
+
+  class Info(
+    msg: Message,
+    pos: SourcePosition
+  ) extends Diagnostic(msg, pos, INFO)
+
+  abstract class ConditionalWarning(
+    msg: Message,
+    pos: SourcePosition
+  ) extends Warning(msg, pos) {
+    def enablingOption(implicit ctx: Context): Setting[Boolean]
+  }
+
+  class FeatureWarning(
+    msg: Message,
+    pos: SourcePosition
+  ) extends ConditionalWarning(msg, pos) {
+    def enablingOption(implicit ctx: Context): Setting[Boolean] = ctx.settings.feature
+  }
+
+  class UncheckedWarning(
+    msg: Message,
+    pos: SourcePosition
+  ) extends ConditionalWarning(msg, pos) {
+    def enablingOption(implicit ctx: Context): Setting[Boolean] = ctx.settings.unchecked
+  }
+
+  class DeprecationWarning(
+    msg: Message,
+    pos: SourcePosition
+  ) extends ConditionalWarning(msg, pos) {
+    def enablingOption(implicit ctx: Context): Setting[Boolean] = ctx.settings.deprecation
+  }
+
+  class MigrationWarning(
+    msg: Message,
+    pos: SourcePosition
+  ) extends ConditionalWarning(msg, pos) {
+    def enablingOption(implicit ctx: Context): Setting[Boolean] = ctx.settings.migration
+  }
 
 class Diagnostic(
   val contained: Message,
