+++ compiler/src/dotty/tools/dotc/typer/Inliner.scala
@@ -222,7 +222,7 @@ object Inliner {
   }
 
   object Intrinsics {
-    import dotty.tools.dotc.reporting.diagnostic.messages.Error
+    import dotty.tools.dotc.reporting.diagnostic.Diagnostic.Error
     private enum ErrorKind:
       case Parser, Typer
